api update

+++ SquirrelScriptTests.cpp
@@ -1149,9 +1149,9 @@ TEST_CASE_FIXTURE(SquirrelScriptTests, "Script.Squirrel.Overloads")
     SUBCASE("Member functions")
     {
         TestScriptClass t;
-        auto            f1 = resolve_overload<i32, f32>(&TestScriptClass::overload);
-        auto            f2 = resolve_overload<f32, i32>(&TestScriptClass::overload);
-        auto            f3 = resolve_overload<>(&TestScriptClass::overload);
+        auto            f1 = resolve_overload<f32(i32, f32)>(&TestScriptClass::overload);
+        auto            f2 = resolve_overload<f32(f32, i32)>(&TestScriptClass::overload);
+        auto            f3 = resolve_overload<f32()>(&TestScriptClass::overload);
 
         auto overload      = make_unique_overload(f1, f2, f3);
         global["obj"]      = &t;
